feat(native): Jinja - support fields reflection for PyObject
chore: update

tmp
