+++ packages/cubejs-backend-native/src/template/mod.rs
@@ -1,4 +1,5 @@
 mod entry;
 mod mj_value;
+mod neon;
 
 pub use entry::template_register_module;
