+++ packages/cubejs-backend-native/test/templates/jinja-instance.py
@@ -49,6 +49,15 @@ def new_str_tuple():
 def new_safe_string():
   return SafeString('"safe string" <>')
 
+class MyCustomObject(dict):
+  def __init__(self):
+    self['a_attr'] = "value for attribute a"
+    self['b_attr'] = "value for attribute b"
+
+@template.function
+def new_object_from_dict():
+  return MyCustomObject()
+
 @template.function
 def load_data_sync():
   client = MyApiClient("google.com")
