+++ packages/cubejs-backend-native/test/templates/scoped-utils.py
@@ -41,6 +41,15 @@ def new_str_tuple():
 def new_safe_string():
   return SafeString('"safe string" <>')
 
+class MyCustomObject(dict):
+  def __init__(self):
+    self['a_attr'] = "value for attribute a"
+    self['b_attr'] = "value for attribute b"
+
+@context_func
+def new_object_from_dict():
+  return MyCustomObject()
+
 @context_func
 def load_data_sync():
    client = MyApiClient("google.com")
